feat: Remove cozy-codeshifts and reorganize code and APIs
BREAKING CHANGE

@cozy/codemods and cozy-codeshifts were quasi duplicates. Here all the
files from cozy-codeshifts are moved to @cozy/codemods and utility files
from @cozy/codemods are organized into concerns:

- hoc.js: deal with higher order components
- imports.js: deal with imports
- jsx.js: deal with jsx constructs
- nodes.js: deal with jsx nodes
etc...
