+++ packages/cozy-codemods/CHANGELOG.md
@@ -3,57 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.2.2](https://github.com/cozy/cozy-libs/compare/@cozy/codemods@1.2.1...@cozy/codemods@1.2.2) (2020-10-07)
-
-**Note:** Version bump only for package @cozy/codemods
-
-
-
-
-
-## [1.2.1](https://github.com/cozy/cozy-libs/compare/@cozy/codemods@1.2.0...@cozy/codemods@1.2.1) (2020-10-01)
-
-**Note:** Version bump only for package @cozy/codemods
-
-
-
-
-
-# [1.2.0](https://github.com/cozy/cozy-libs/compare/@cozy/codemods@1.1.2...@cozy/codemods@1.2.0) (2020-09-14)
+# 1.4.0 (2019-09-05)
 
 
 ### Features
 
-* Add helper functions to transform imports ([807ff9f](https://github.com/cozy/cozy-libs/commit/807ff9f))
+* Add account route ([7986708](https://github.com/cozy/cozy-libs/commit/7986708))
 
 
 
 
 
-## [1.1.2](https://github.com/cozy/cozy-libs/compare/@cozy/codemods@1.1.1...@cozy/codemods@1.1.2) (2020-09-01)
-
-**Note:** Version bump only for package @cozy/codemods
-
-
-
-
-
-## [1.1.1](https://github.com/cozy/cozy-libs/compare/@cozy/codemods@1.1.0...@cozy/codemods@1.1.1) (2020-08-03)
-
-**Note:** Version bump only for package @cozy/codemods
-
-
-
-
-
-# 1.1.0 (2020-05-11)
-
-
-### Bug Fixes
-
-* Ability to process HOC without options ([01d176e](https://github.com/cozy/cozy-libs/commit/01d176e))
+# [1.3.0](https://github.com/cozy/cozy-libs/compare/cozy-codeshifts@1.2.0...cozy-codeshifts@1.3.0) (2019-07-18)
 
 
 ### Features
 
-* Add codemods shared package ([1fa199e](https://github.com/cozy/cozy-libs/commit/1fa199e))
+* Add remove unused imports transform ([4e2d715](https://github.com/cozy/cozy-libs/commit/4e2d715))
+* Add the add-import util ([2de9792](https://github.com/cozy/cozy-libs/commit/2de9792))
+
+
+
+
+
+# [1.2.0](https://github.com/cozy/cozy-libs/compare/cozy-codeshifts@1.1.0...cozy-codeshifts@1.2.0) (2019-07-05)
+
+
+### Features
+
+* Add remove redundant expression containers ([3e7b08a](https://github.com/cozy/cozy-libs/commit/3e7b08a))
+* Remove redundant expression containers. Added test ([ad027a8](https://github.com/cozy/cozy-libs/commit/ad027a8))
+
+
+
+
+
+# 1.1.0 (2019-07-04)
+
+
+### Features
+
+* Add codeshift transforms + utils ([f212ec6](https://github.com/cozy/cozy-libs/commit/f212ec6))
