+++ packages/cozy-codemods/README.md
@@ -1,6 +1,148 @@
-### Cozy codemods
+<!-- Autogenerated, change .README.template -->
 
-Helpers to create codemods with [`jscodeshift`](https://github.com/facebook/jscodeshift).
+## Cozy codeshifts
+
+A collection of utils and transforms for [jscodeshift].
+
+Codeshifts are automatic transformations of Javascript code. They can be used
+for
+
+- code migrations following API changes
+- flag removal
+- general automatic code changes...
+
+See [here][jscodeshift-docs] for more information on codeshifts.
+
+[jscodeshift]: https://github.com/facebook/jscodeshift
+[jscodeshift-docs]: https://github.com/facebook/jscodeshift/wiki/jscodeshift-Documentation
+
+### General usage
+
+```
+$ jscodeshift -t node_modules/cozy-codeshifts/dist/<TRANSFORM> --parser babel --js,jsx -- src
+```
+
+### Available transforms
+
+#### Apply flag
+
+Transforms `flag()` calls into `true` then remove dead code and dead imports
+
+Before
+
+```
+import Old from 'old'
+import New from 'new'
+flag("hello") ? <Old /> : <New />
+```
+
+After
+
+```
+import New from 'new'
+<New />
+```
+
+#### Remove boolean variables
+
+Simpler version of Apply flag only for boolean variables.
+
+Before
+
+```
+import Old from 'old'
+import New from 'new'
+true ? <Old /> : <New />
+```
+
+After
+
+```
+<New />
+```
+
+### Utils
 
 - Replace a HOC with a hook with [`hoc-replacer.js`](./src/hoc-replacer.js)
 - Remove unused imports with [`remove-unused-imports.js`](./src/remove-unused-imports.js)
+
+
+## API
+
+## Functions
+
+<dl>
+<dt><a href="#removeUnusedImports">removeUnusedImports(root, j)</a></dt>
+<dd><p>Removes unused imports by counting usage.
+JSX identifiers are counted as React usage.</p>
+</dd>
+<dt><a href="#isBlockLike">isBlockLike(path)</a> ⇒ <code>Boolean</code></dt>
+<dd><p>Returns true if path is Program or a Block</p>
+</dd>
+<dt><a href="#flatReplace">flatReplace(path, newNode)</a></dt>
+<dd><p>Replaces <code>path.node</code> with <code>newNode</code> without keeping blocks, flattening
+<code>newNode</code> into <code>path</code>. Useful when removing <code>if</code>/<code>else</code>.</p>
+</dd>
+<dt><a href="#simplifyConditions">simplifyConditions(root, j)</a></dt>
+<dd><p>Statically evaluates boolean conditions</p>
+</dd>
+</dl>
+
+<a name="removeUnusedImports"></a>
+
+## removeUnusedImports(root, j)
+Removes unused imports by counting usage.
+JSX identifiers are counted as React usage.
+
+**Kind**: global function
+
+| Param | Type |
+| --- | --- |
+| root | <code>PathNode</code> |
+| j | <code>Object</code> |
+
+<a name="isBlockLike"></a>
+
+## isBlockLike(path) ⇒ <code>Boolean</code>
+Returns true if path is Program or a Block
+
+**Kind**: global function
+
+| Param | Type |
+| --- | --- |
+| path | <code>PathNode</code> |
+
+<a name="flatReplace"></a>
+
+## flatReplace(path, newNode)
+Replaces `path.node` with `newNode` without keeping blocks, flattening
+`newNode` into `path`. Useful when removing `if`/`else`.
+
+**Kind**: global function
+
+| Param | Type |
+| --- | --- |
+| path | <code>PathNode</code> |
+| newNode | <code>Node</code> |
+
+<a name="simplifyConditions"></a>
+
+## simplifyConditions(root, j)
+Statically evaluates boolean conditions
+
+**Kind**: global function
+
+| Param | Type |
+| --- | --- |
+| root | <code>NodePath</code> |
+| j | <code>Object</code> |
+
+**Example**
+```js
+`if (true) { foo } else { bar }` -> `foo`
+
+`true ? foo : bar` -> `foo`
+
+`!true ? foo : bar` -> `bar`
+```
+
