+++ packages/cozy-codemods/src/index.js
@@ -1,5 +1,16 @@
-const makeUtils = require('./utils')
-const findNearest = require('./find-nearest')
-const hocReplacer = require('./hoc-replacer')
+const hoc = require('./hoc')
+const imports = require('./imports')
+const jsx = require('./jsx')
 
-module.exports = { makeUtils, findNearest, hocReplacer }
+const replaceBooleanVars = require('./replace-boolean-vars')
+const simplifyConditions = require('./simplify-conditions')
+const findNearest = require('./find-nearest')
+
+module.exports = {
+  hoc,
+  imports,
+  jsx,
+  replaceBooleanVars,
+  simplifyConditions,
+  findNearest
+}
