+++ packages/cozy-codemods/src/transforms/__testfixtures__/apply-flag.input.js
