+++ packages/cozy-codemods/src/transforms/__testfixtures__/apply-flag.output.js
