+++ packages/cozy-codemods/src/transforms/__tests__/apply-flag.spec.js
