--- packages/cozy-codeshifts/CHANGELOG.md
@@ -1,46 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-# 1.4.0 (2019-09-05)
-
-
-### Features
-
-* Add account route ([7986708](https://github.com/cozy/cozy-libs/commit/7986708))
-
-
-
-
-
-# [1.3.0](https://github.com/cozy/cozy-libs/compare/cozy-codeshifts@1.2.0...cozy-codeshifts@1.3.0) (2019-07-18)
-
-
-### Features
-
-* Add remove unused imports transform ([4e2d715](https://github.com/cozy/cozy-libs/commit/4e2d715))
-* Add the add-import util ([2de9792](https://github.com/cozy/cozy-libs/commit/2de9792))
-
-
-
-
-
-# [1.2.0](https://github.com/cozy/cozy-libs/compare/cozy-codeshifts@1.1.0...cozy-codeshifts@1.2.0) (2019-07-05)
-
-
-### Features
-
-* Add remove redundant expression containers ([3e7b08a](https://github.com/cozy/cozy-libs/commit/3e7b08a))
-* Remove redundant expression containers. Added test ([ad027a8](https://github.com/cozy/cozy-libs/commit/ad027a8))
-
-
-
-
-
-# 1.1.0 (2019-07-04)
-
-
-### Features
-
-* Add codeshift transforms + utils ([f212ec6](https://github.com/cozy/cozy-libs/commit/f212ec6))
