--- packages/cozy-codeshifts/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "cozy-codeshifts",
-  "version": "1.4.0",
-  "description": "Codeshifts used inside Cozies",
-  "main": "index.js",
-  "repository": "https://github.com/cozy/cozy-libs",
-  "author": "Cozy",
-  "license": "MIT",
-  "scripts": {
-    "docs": "jsdoc2md -f 'src/**/*.js' -t .README.md.template > README.md",
-    "build": "babel src -d dist --verbose"
-  },
-  "devDependencies": {
-    "jest": "^24.8.0",
-    "jscodeshift": "^0.6.4",
-    "jsdoc-to-markdown": "^5.0.0"
-  }
-}
